Genericise DownloadManager, add failure callback, create interfaces and delegates. Add ability to skip re-queueing
