+++ GChan/GChan.csproj
@@ -80,7 +80,7 @@
     <Reference Include="System.Xml.Serialization" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Controllers\FileDownloadController.cs" />
+    <Compile Include="Controllers\DownloadManager.cs" />
     <Compile Include="Data\DataController.cs" />
     <Compile Include="Controllers\MainController.cs" />
     <Compile Include="Controllers\UpdateController.cs" />
@@ -94,7 +94,7 @@
     <Compile Include="Forms\AboutBox.Designer.cs">
       <DependentUpon>AboutBox.cs</DependentUpon>
     </Compile>
-    <Compile Include="EnumHelper.cs" />
+    <Compile Include="Helpers\EnumHelper.cs" />
     <Compile Include="Forms\MainForm.cs">
       <SubType>Form</SubType>
     </Compile>
@@ -109,7 +109,8 @@
     </Compile>
     <Compile Include="Helpers\EnumerableExtensions.cs" />
     <Compile Include="Helpers\JTokenExtensions.cs" />
-    <Compile Include="ImageLink.cs" />
+    <Compile Include="Trackers\IDownloadable.cs" />
+    <Compile Include="Trackers\ImageLink.cs" />
     <Compile Include="Controls\PreferencesDataGridView.cs">
       <SubType>Component</SubType>
     </Compile>
@@ -134,7 +135,7 @@
     <Compile Include="Trackers\Thread.cs" />
     <Compile Include="Trackers\Sites\Thread_4Chan.cs" />
     <Compile Include="Trackers\Tracker.cs" />
-    <Compile Include="Utils.cs" />
+    <Compile Include="Helpers\Utils.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Controls\SortableBindingList\PropertyComparer.cs" />
