+++ GChan/Trackers/IDownloadable.cs
@@ -1,0 +1,12 @@
+﻿using GChan.Controllers;
+
+namespace GChan
+{
+    public interface IDownloadable<T> where T : IDownloadable<T>
+    {
+        void Download(
+            DownloadManager<T>.SuccessCallback successCallback,
+            DownloadManager<T>.FailureCallback failureCallback
+        );
+    }
+}
